+++ src/constants.js
@@ -1,7 +1,6 @@
 // Import Third-party Dependencies
 import kleur from "kleur";
 import isUnicodeSupported from "is-unicode-supported";
-import isCI from "is-ci";
 
 const kMainSymbols = {
   tick: "✔",
@@ -21,7 +20,7 @@ const kFallbackSymbols = {
   active: "(+)",
   inactive: "(-)"
 };
-const kSymbols = isUnicodeSupported() || isCI ? kMainSymbols : kFallbackSymbols;
+const kSymbols = isUnicodeSupported() || process.env.CI ? kMainSymbols : kFallbackSymbols;
 const kPointer = kleur.gray(kSymbols.pointer);
 
 export const SYMBOLS = {
