+++ src/prompts/index.js
@@ -1,0 +1,12 @@
+// Import Internal Dependencies
+import { QuestionPrompt } from "./question.js";
+import { ConfirmPrompt } from "./confirm.js";
+import { SelectPrompt } from "./select.js";
+import { MultiselectPrompt } from "./multiselect.js";
+
+export {
+  QuestionPrompt,
+  ConfirmPrompt,
+  SelectPrompt,
+  MultiselectPrompt
+};
