+++ src/prompts/multiselect.js
@@ -7,13 +7,17 @@ import stripAnsi from "strip-ansi";
 import wcwidth from "@topcli/wcwidth";
 
 // Import Internal Dependencies
-import { AbstractPrompt } from "./abstract-prompt.js";
-import { SYMBOLS } from "./constants.js";
+import { AbstractPrompt } from "./abstract.js";
+import { SYMBOLS } from "../constants.js";
 
-export class SelectPrompt extends AbstractPrompt {
+export class MultiselectPrompt extends AbstractPrompt {
   #boundExitEvent = () => void 0;
   #boundKeyPressEvent = () => void 0;
+  #validators;
+
   activeIndex = 0;
+  selectedIndexes = [];
+  questionMessage;
 
   get choices() {
     return this.options.choices;
@@ -23,7 +27,9 @@ export class SelectPrompt extends AbstractPrompt {
     const {
       stdin = process.stdin,
       stdout = process.stdout,
-      choices
+      choices,
+      preSelectedChoices,
+      validators = []
     } = options ?? {};
 
     super(message, stdin, stdout);
@@ -56,6 +62,28 @@ export class SelectPrompt extends AbstractPrompt {
 
       return choice.label.length;
     }));
+
+    this.#validators = validators;
+
+    if (!preSelectedChoices) {
+      return;
+    }
+
+    for (const choice of preSelectedChoices) {
+      const choiceIndex = this.choices.findIndex((item) => {
+        if (typeof item === "string") {
+          return item === choice;
+        }
+
+        return item.value === choice;
+      });
+
+      if (choiceIndex === -1) {
+        throw new Error(`Invalid pre-selected choice: ${choice.value ?? choice}`);
+      }
+
+      this.selectedIndexes.push(choiceIndex);
+    }
   }
 
   #getFormattedChoice(choiceIndex) {
@@ -86,35 +114,37 @@ export class SelectPrompt extends AbstractPrompt {
 
     for (let choiceIndex = startIndex; choiceIndex < endIndex; choiceIndex++) {
       const choice = this.#getFormattedChoice(choiceIndex);
-      const isChoiceSelected = choiceIndex === this.activeIndex;
+      const isChoiceActive = choiceIndex === this.activeIndex;
+      const isChoiceSelected = this.selectedIndexes.includes(choiceIndex);
       const showPreviousChoicesArrow = startIndex > 0 && choiceIndex === startIndex;
       const showNextChoicesArrow = endIndex < this.choices.length && choiceIndex === endIndex - 1;
 
-      let prefixArrow = " ";
+      let prefixArrow = "  ";
       if (showPreviousChoicesArrow) {
-        prefixArrow = SYMBOLS.Previous;
+        prefixArrow = SYMBOLS.Previous + " ";
       }
       else if (showNextChoicesArrow) {
-        prefixArrow = SYMBOLS.Next;
+        prefixArrow = SYMBOLS.Next + " ";
       }
 
-      const prefix = `${prefixArrow}${isChoiceSelected ? `${SYMBOLS.Pointer} ` : "  "}`;
+      const prefix = `${prefixArrow}${isChoiceSelected ? SYMBOLS.Active : SYMBOLS.Inactive}`;
       const formattedLabel = choice.label.padEnd(
         this.longestChoice < 10 ? this.longestChoice : 0
       );
       const formattedDescription = choice.description ? ` - ${choice.description}` : "";
-      const color = isChoiceSelected ? kleur.white().bold : kleur.gray;
-      const str = color(`${prefix}${formattedLabel}${formattedDescription}${EOL}`);
+      const color = isChoiceActive ? kleur.white().bold : kleur.gray;
+      const str = `${prefix} ${color(`${formattedLabel}${formattedDescription}`)}${EOL}`;
 
       this.write(str);
     }
   }
 
-  #showAnsweredQuestion(choice) {
-    const prefix = `${SYMBOLS.Tick} ${kleur.bold(this.message)} ${SYMBOLS.Pointer}`;
-    const formattedChoice = kleur.yellow(choice.label ?? choice);
+  #showAnsweredQuestion(choices, isAgentAnswer = false) {
+    const prefixSymbol = this.selectedIndexes.length === 0 && !isAgentAnswer ? SYMBOLS.Cross : SYMBOLS.Tick;
+    const prefix = `${prefixSymbol} ${kleur.bold(this.message)} ${SYMBOLS.Pointer}`;
+    const formattedChoice = kleur.yellow(choices);
 
-    this.write(`${prefix} ${formattedChoice}${EOL}`);
+    this.write(`${prefix}${choices ? ` ${formattedChoice}` : ""}${EOL}`);
   }
 
   #onProcessExit() {
@@ -135,32 +165,56 @@ export class SelectPrompt extends AbstractPrompt {
       this.activeIndex = this.activeIndex === this.choices.length - 1 ? 0 : this.activeIndex + 1;
       render();
     }
+    else if (key.name === "a") {
+      this.selectedIndexes = this.selectedIndexes.length === this.choices.length ? [] : this.choices.map((_, index) => index);
+      render();
+    }
+    else if (key.name === "space") {
+      const isChoiceSelected = this.selectedIndexes.includes(this.activeIndex);
+
+      if (isChoiceSelected) {
+        this.selectedIndexes = this.selectedIndexes.filter((index) => index !== this.activeIndex);
+      }
+      else {
+        this.selectedIndexes.push(this.activeIndex);
+      }
+
+      render();
+    }
     else if (key.name === "return") {
+      const labels = this.selectedIndexes.map((index) => this.choices[index].label ?? this.choices[index]);
+      const values = this.selectedIndexes.map((index) => this.choices[index].value ?? this.choices[index]);
+
+      for (const validator of this.#validators) {
+        if (!validator.validate(values)) {
+          const error = validator.error(values);
+          render({ error });
+
+          return;
+        }
+      }
+
       render({ clearRender: true });
 
-      const currentChoice = this.choices[this.activeIndex];
-      const value = currentChoice.value ?? currentChoice;
-
-      if (!this.options.ignoreValues?.includes(value)) {
-        this.#showAnsweredQuestion(currentChoice);
-      }
+      this.#showAnsweredQuestion(labels.join(", "));
 
       this.write(SYMBOLS.ShowCursor);
       this.destroy();
 
       this.#onProcessExit();
       process.off("exit", this.#boundExitEvent);
-      resolve(value);
+
+      resolve(values);
     }
     else {
       render();
     }
   }
 
-  async select() {
+  async multiselect() {
     if (this.agent.nextAnswers.length > 0) {
       const answer = this.agent.nextAnswers.shift();
-      this.#showAnsweredQuestion(answer);
+      this.#showAnsweredQuestion(answer, true);
       this.destroy();
 
       return answer;
@@ -172,7 +226,8 @@ export class SelectPrompt extends AbstractPrompt {
     const render = (options = {}) => {
       const {
         initialRender = false,
-        clearRender = false
+        clearRender = false,
+        error = null
       } = options;
 
       if (!initialRender) {
@@ -187,11 +242,16 @@ export class SelectPrompt extends AbstractPrompt {
         const questionLineCount = Math.ceil(
           wcwidth(stripAnsi(this.questionMessage)) / this.stdout.columns
         );
-
         this.stdout.moveCursor(-this.stdout.columns, -(1 + questionLineCount));
         this.stdout.clearScreenDown();
 
         return;
+      }
+
+      if (error) {
+        this.stdout.moveCursor(0, -2);
+        this.stdout.clearScreenDown();
+        this.#showQuestion(error);
       }
 
       this.#showChoices();
@@ -208,8 +268,15 @@ export class SelectPrompt extends AbstractPrompt {
     });
   }
 
-  #showQuestion() {
-    this.questionMessage = `${SYMBOLS.QuestionMark} ${kleur.bold(this.message)}`;
+  #showQuestion(error = null) {
+    let hint = kleur.gray(
+      `(Press ${kleur.bold("<a>")} to toggle all, ${kleur.bold("<space>")} to select, ${kleur.bold("<return>")} to submit)`
+    );
+    if (error) {
+      hint += ` ${kleur.red().bold(`[${error}]`)}`;
+    }
+    this.questionMessage = `${SYMBOLS.QuestionMark} ${kleur.bold(this.message)} ${hint}`;
+
     this.write(`${this.questionMessage}${EOL}`);
   }
 }
